:truck: rename: Google API 별로 디렉토리 구분
