+++ src/main/java/zzandori/zzanmoa/googleapi/dto/findplace/FindPlaceResponse.java
@@ -1,4 +1,4 @@
-package zzandori.zzanmoa.googleapi.dto;
+package zzandori.zzanmoa.googleapi.dto.findplace;
 
 import java.util.List;
 import lombok.Getter;
