+++ src/main/java/zzandori/zzanmoa/googleapi/dto/geometry/GeocodeRequest.java
@@ -1,4 +1,4 @@
-package zzandori.zzanmoa.googleapi.dto;
+package zzandori.zzanmoa.googleapi.dto.geometry;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -7,21 +7,18 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class FindPlaceRequest {
+public class GeocodeRequest {
 
-    @Value("${GOOGLE_MAP_API_FINDPLACE_BASE_URL}")
-    private String GOOGLE_MAP_API_FINDPLACE_BASE_URL;
+    @Value("${GOOGLE_MAP_API_GEOCODE_BASE_URL}")
+    private String GOOGLE_MAP_API_GEOCODE_BASE_URL;
     @Value("${GOOGLE_MAP_API_KEY}")
     private String GOOGLE_MAP_API_KEY;
 
     public String requestUrl(String address) throws UnsupportedEncodingException {
         String encodedAddress = URLEncoder.encode(address, StandardCharsets.UTF_8.toString());
-        StringBuilder urlBuilder = new StringBuilder(GOOGLE_MAP_API_FINDPLACE_BASE_URL)
-            .append("?input=").append(encodedAddress)
-            .append("&inputtype=").append("textquery")
-            .append("&fields=").append("formatted_address").append("%2C").append("place_id")
+        StringBuilder urlBuilder = new StringBuilder(GOOGLE_MAP_API_GEOCODE_BASE_URL)
+            .append("?address=").append(encodedAddress)
             .append("&key=").append(GOOGLE_MAP_API_KEY);
         return urlBuilder.toString();
     }
-
 }
