+++ src/main/java/zzandori/zzanmoa/googleapi/dto/geometry/Result.java
@@ -1,4 +1,4 @@
-package zzandori.zzanmoa.googleapi.dto;
+package zzandori.zzanmoa.googleapi.dto.geometry;
 
 import lombok.Getter;
